Cross exp phase3

+++ hits_cpp/types.hpp
@@ -14,7 +14,7 @@ typedef map<string, list<string>*> SequencesMap;
 typedef map<char, int> AlphabetMap;
 typedef map<string, double> CutoffsMap;
 typedef vector<vector<double>> MemeRows;
-typedef map<string, int> SequencesCount;
+typedef map<string, float> SequencesCount;
 typedef vector<int> ShufflePattern;
 typedef vector<ShufflePattern> ShufflePatterns;
 typedef map<int, ShufflePatterns> ShufflesMap;
